feat(material_picker.dart): Add support for custom color types
This commit introduces the ability to customize the color types and initial color type in the MaterialPicker widget.  The `colorTypes` and `initialColorType` parameters allow developers to provide their own lists of colors, offering greater flexibility and control over the picker's appearance and functionality.  Default color types are still used if custom types are not provided.